requirements.txt
streamlit==1.33.0
google-cloud-speech
google-cloud-aiplatform
vertexai
streamlit-webrtc
